image-retrieval-api/requirements.txt
fastapi
uvicorn
numpy
requests
httpx
pillow
tensorflow
